Add files via upload
I enjoyed so much while working on this project due to my interest in cricket. In this project i have done complete analysis of 2019 cricket world cup involving batting, bowling analysis and predicting the best batsman and bowlers. Team analysis and the Ground analysis is also done for all the hosting venues of England.